Don't cut the url on a redirect
This is a fix for issue 2485033. It is not a final fix, but just starting with this so it can be discussed on code review. The line in question
was added to fix issue 1690652.

Updated fix. Stores the url before it is cut during a redirect. Forwards this to the reponse instead of the cut url.

Update 2: Using the old originalUrl

Change-Id: I286084451aa45e51d5d07811f9d119cf83849592