Make HiPri work when already connected.
A recent change made the HiPri MobileDataStateTracker listen for notifications
about the default connection (which HiPri shadows).  Local code was sending
itself a notification using the old HiPri badging instead of the new Default
badging and those notifications where therefore ignored.

Manifested itself on HiPri connections when we were already on 3g.

See change 42422 on master platform/frameworks/base for the change this is completing.

Change-Id: I375026048724d0035297287c61c6c2f58d4e0294