Stupid AutoCompleteTextView. Bug #2522538
Fixes problems that occur in the Email application. If the adapter is initially empty,
the code in the data set observer to update the popup would not show the popup. This
change makes sure the popup will be shown if the adapter was initially empty and the
user has typed enough characters in the input field.

Change-Id: I44a0e4fab18a642763816a974b8c1886d8e52869