Fix 2537999: Don't throw exception when SlidingTab has UNSPECIFIED specs
This is try to capture the case where we sometimes see UNSPECIFIED
MeasureSpec in SlidingTab. It should be OK as long as one of the dimensions
is constrained. The log event will give some indication of what's
going on when this happens.

This also fixes a bug where it was possible to not reset the mAnimating flag
in the Phone app.

Change-Id: Ibab1c944ec895a7c19d5a18cc67f6e4db716f685