Enable contentEditable.
WebView.java:
If the cursor and focus are on a contentEditable element, allow it
to handle navigation keys and shift rather than the navigation system.
Also rename method for its new more general use.

WebViewCore.java:
Change the name of MOVE_OUT_OF_PLUGIN to a more generic name, and
fix a bug in the message being sent.

Bug 1788820

Requires a change to external/webkit

Caveats:
Does not ensure that the caret remains on screen.  Frame::revealSelection
is called, but we ignore it for other reasons.  Need to investigate that.
The cursor will blink if the contentEditable node has focus, even if the
user has not clicked on it or has moved to a different input field. Further,
while in this state, the user can input text.

Change-Id: Iaa935363fd324925a1e338418d32f362f58c4494